# src/swagger_files/QuickCode.Turuncu.UserManagerModule.Api-Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickCode - Turuncu - UserManagerModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/user-manager-module/asp-net-users/{aspNetUsersId}/refresh-tokens/{refreshTokensId}": {
      "get": {
        "tags": [
          "AspNetUsers"
        ],
        "operationId": "api-user-manager-module-asp-net-users--aspNetUsersId--refresh-tokens--refreshTokensId--get",
        "parameters": [
          {
            "name": "aspNetUsersId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refreshTokensId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AspNetUsersRefreshTokens_KEY_RESTResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/user-manager-module/refresh-tokens/count": {
      "get": {
        "tags": [
          "RefreshTokens"
        ],
        "operationId": "api-user-manager-module-refresh-tokens-count-get",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/refresh-tokens/{id}": {
      "get": {
        "tags": [
          "RefreshTokens"
        ],
        "operationId": "api-user-manager-module-refresh-tokens--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokensDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RefreshTokens"
        ],
        "operationId": "api-user-manager-module-refresh-tokens--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokensDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokensDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokensDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RefreshTokens"
        ],
        "operationId": "api-user-manager-module-refresh-tokens--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/refresh-tokens/get-refresh-token/{refreshTokensToken}": {
      "get": {
        "tags": [
          "RefreshTokens"
        ],
        "operationId": "api-user-manager-module-refresh-tokens-get-refresh-token--refreshTokensToken--get",
        "parameters": [
          {
            "name": "refreshTokensToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokensGetRefreshTokenResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/user-manager-module/refresh-tokens/update-refresh-token/{refreshTokensToken}": {
      "patch": {
        "tags": [
          "RefreshTokens"
        ],
        "operationId": "api-user-manager-module-refresh-tokens-update-refresh-token--refreshTokensToken--patch",
        "parameters": [
          {
            "name": "refreshTokensToken",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokensUpdateRefreshTokenRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokensUpdateRefreshTokenRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokensUpdateRefreshTokenRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AspNetUsersRefreshTokens_KEY_RESTResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "expiryDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "isRevoked": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "RefreshTokensDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "expiryDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "isRevoked": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RefreshTokensGetRefreshTokenResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "expiryDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "isRevoked": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "RefreshTokensUpdateRefreshTokenRequestDto": {
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}